[REFACTOR] 결제 도메인 삭제 및 결제 호출을 reservation으로 이동
